+++ functions/clocks.py
@@ -14,7 +14,7 @@ import time
 from rpi_ws281x import *
 
 from control.light_wreath import wreath_setup
-from functions.effect import clear, wipe_second
+from functions.effects import clear, wipe_second
 from logger import LOGGER
